Make pages dir structure match nav structure
